requirements.txt
python_version < 3.12
discord.py==2.3.2
openai>=1.0.0
langchain==0.0.325
requests
python-dotenv
tweepy
flask
